multithreaded multi-lookahead works

+++ core/src/gridparts/GfxState.java
@@ -25,7 +25,7 @@ public class GfxState {
     public float               redHandYOffset;
     public float               blueHandYOffset;
     
-    public int                 thinkingReady;
+    public double              thinkingReady;
     public double              thinkingTotal;
     
     private Menu               menu;
